Test for statistics netvalue REST API endpoint
