+++ rotkehlchen/tests/utils/rotkehlchen.py
@@ -1,9 +1,78 @@
-from typing import List
+from typing import Dict, List, NamedTuple
+from unittest.mock import _patch
+
+import requests
 
 from rotkehlchen.constants.assets import A_BTC, A_ETH, A_EUR
 from rotkehlchen.db.utils import AssetBalance, LocationData
+from rotkehlchen.fval import FVal
+from rotkehlchen.tests.utils.blockchain import mock_etherscan_balances_query
 from rotkehlchen.tests.utils.constants import A_XMR
-from rotkehlchen.typing import Location, Timestamp
+from rotkehlchen.tests.utils.exchanges import (
+    patch_binance_balances_query,
+    patch_poloniex_balances_query,
+)
+from rotkehlchen.typing import BTCAddress, ChecksumEthAddress, Location, Timestamp
+
+
+class BalancesTestSetup(NamedTuple):
+    eth_balances: List[ChecksumEthAddress]
+    btc_balances: List[BTCAddress]
+    eur_balance: FVal
+    rdn_balance: FVal
+    binance_balances: Dict[str, FVal]
+    poloniex_balances: Dict[str, FVal]
+    poloniex_patch: _patch
+    binance_patch: _patch
+    blockchain_patch: _patch
+
+
+def setup_balances(
+        rotki,
+        ethereum_accounts: List[ChecksumEthAddress],
+        btc_accounts: List[BTCAddress],
+) -> BalancesTestSetup:
+    """Setup the blockchain, exchange and fiat balances for some tests"""
+    eth_acc1 = ethereum_accounts[0]
+    eth_acc2 = ethereum_accounts[1]
+    eth_balance1 = '1000000'
+    eth_balance2 = '2000000'
+    eth_balances = [eth_balance1, eth_balance2]
+    btc_balance1 = '3000000'
+    btc_balance2 = '5000000'
+    btc_balances = [btc_balance1, btc_balance2]
+    rdn_balance = '4000000'
+    eur_balance = FVal('1550')
+
+    rotki.data.db.add_fiat_balance(A_EUR, eur_balance)
+    binance = rotki.exchange_manager.connected_exchanges['binance']
+    poloniex = rotki.exchange_manager.connected_exchanges['poloniex']
+    poloniex_patch = patch_poloniex_balances_query(poloniex)
+    binance_patch = patch_binance_balances_query(binance)
+    blockchain_patch = mock_etherscan_balances_query(
+        eth_map={
+            eth_acc1: {'ETH': eth_balance1},
+            eth_acc2: {'ETH': eth_balance2, 'RDN': rdn_balance},
+        },
+        btc_map={btc_accounts[0]: btc_balance1, btc_accounts[1]: btc_balance2},
+        original_requests_get=requests.get,
+    )
+    # Taken from BINANCE_BALANCES_RESPONSE from tests.utils.exchanges
+    binance_balances = {'ETH': FVal('4763368.68006011'), 'BTC': FVal('4723846.89208129')}
+    # Taken from POLONIEX_BALANCES_RESPONSE from tests.utils.exchanges
+    poloniex_balances = {'ETH': FVal('11.0'), 'BTC': FVal('5.5')}
+
+    return BalancesTestSetup(
+        eth_balances=eth_balances,
+        btc_balances=btc_balances,
+        rdn_balance=rdn_balance,
+        eur_balance=eur_balance,
+        binance_balances=binance_balances,
+        poloniex_balances=poloniex_balances,
+        poloniex_patch=poloniex_patch,
+        binance_patch=binance_patch,
+        blockchain_patch=blockchain_patch,
+    )
 
 
 def add_starting_balances(datahandler) -> List[AssetBalance]:
